+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "nsls2ptycho/core/ptycho"]
-	path = nsls2ptycho/core/ptycho
+[submodule "nsls2ptycho/core/backends/ptycho"]
+	path = nsls2ptycho/core/backends/ptycho
 	url = ../ptycho.git
