+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include nsls2ptycho/core/ptycho/*.cubin
+include nsls2ptycho/core/backends/ptycho/*.cubin
