+++ nsls2ptycho/core/__init__.py
@@ -1,3 +1,3 @@
+import nsls2ptycho.core.backends.ptycho
 import nsls2ptycho.core.databroker
-import nsls2ptycho.core.ptycho
 import nsls2ptycho.core.widgets
